[lumen-app] More reasonable precision settings in presets configs

+++ lumen-app/src/lumen_app/core/config.py
@@ -329,7 +329,7 @@ class Config:
                             model="PP-OCRv5",
                             runtime=self.unified_runtime,
                             rknn_device=self.unified_rknn_device,
-                            precision="fp16",
+                            precision="fp32",
                             dataset=None,
                         )
                     },
@@ -378,7 +378,7 @@ class Config:
                             model="PP-OCRv5",
                             runtime=self.unified_runtime,
                             rknn_device=self.unified_rknn_device,
-                            precision="fp16",
+                            precision="fp32",
                             dataset=None,
                         )
                     },
@@ -401,7 +401,7 @@ class Config:
                             runtime=self.unified_runtime,
                             rknn_device=self.unified_rknn_device,
                             dataset="ImageNet_1k",
-                            precision=self.device_config.precision or "int8",
+                            precision=self.device_config.precision or "fp16",
                         )
                     },
                 ),
@@ -422,7 +422,7 @@ class Config:
                             model="buffalo_l",
                             runtime=self.unified_runtime,
                             rknn_device=self.unified_rknn_device,
-                            precision=self.device_config.precision or "int8",
+                            precision=self.device_config.precision or "fp16",
                             dataset=None,
                         )
                     },
@@ -495,7 +495,7 @@ class Config:
                             runtime=self.unified_runtime,
                             rknn_device=self.unified_rknn_device,
                             dataset="ImageNet_1k",
-                            precision=self.device_config.precision or "int8",
+                            precision=self.device_config.precision or "fp16",
                         )
                     },
                 ),
@@ -516,7 +516,7 @@ class Config:
                             model="antelopev2",
                             runtime=self.unified_runtime,
                             rknn_device=self.unified_rknn_device,
-                            precision=self.device_config.precision or "int8",
+                            precision=self.device_config.precision or "fp16",
                             dataset=None,
                         )
                     },
@@ -652,7 +652,7 @@ class Config:
                             model="FastVLM-0.5B",
                             runtime=self.unified_runtime,
                             rknn_device=self.unified_rknn_device,
-                            precision=self.device_config.precision or "int8",
+                            precision=self.device_config.precision or "fp16",
                             dataset=None,
                         )
                     },
